+++ scripts/macros/generic/damageBonusToOneRoll.mjs
@@ -1,8 +1,12 @@
 import {automationUtils, constants, DamageBonus, documentUtils, workflowUtils} from '../../proxy.mjs';
 async function damage({document, workflow}) {
-    if (!workflow.targets.size) return;
+    if (!workflow.targets.size || (!workflow.activity.hasDamage && !workflow.activity.hasHealing)) return;
     const config = automationUtils.getGenericConfigValues(document, 'chris-premades', 'damageBonusToOneRoll', Object.keys(damageBonusToOneRoll.genericConfig));
     if (!config.bonus.length) return;
+    const multiSingleTarget = workflow.workflowOptions['chris-premades']?.multiSingleTarget;
+    if (multiSingleTarget && document.flags['chris-premades']?.lastUse === multiSingleTarget.rollID) return;
+    const optional = config.useActivityCosts || (multiSingleTarget ? multiSingleTarget.remainingAttacks > 1 : false);
+    if (optional && workflow.activity.hasAttack && !workflow.hitTargets.size) return;
     if (config.attackType.length) {   
         if (!workflowUtils.isAttackType(workflow, config.attackType)) return;
     }
@@ -22,13 +26,13 @@ async function damage({document, workflow}) {
     if (config.itemType.length) {
         if (!config.itemType.includes(workflow.item.type)) return;
     }
+    if (config.spellLevel.length) {
+        if(!config.spellLevel.includes(String(workflowUtils.getCastLevel(workflow)))) return;
+    }
     if (config.spellSchool.length) {
         if (!config.spellSchool.includes(workflow.item.system.school)) return;
     }
-    const multiSingleTarget = workflow.workflowOptions['chris-premades']?.multiSingleTarget;
-    if (multiSingleTarget && document.flags['chris-premades']?.lastUse === multiSingleTarget.rollID) return;
-    const optional = multiSingleTarget ? multiSingleTarget.remainingAttacks > 1 : false;
-    return new DamageBonus(document, {formula: config.bonus, optional})
+    const bonus = new DamageBonus(document, {formula: config.bonus, optional, type: config.bonusDamageType, allowCritical: config.allowCritical})
         .withOnUse(async ({bonus}) => {
             if (multiSingleTarget) await documentUtils.setFlag(document, 'chris-premades', 'lastUse', multiSingleTarget.rollID);
             const item = bonus.document.documentName === 'Item' ? bonus.document : bonus.activity?.item;
@@ -40,8 +44,14 @@ async function damage({document, workflow}) {
                 if (!activity) return;
                 await workflowUtils.completeActivityUse(activity, Array.from(bonus.targets ?? []));
             }
-        })
-        .initialize();
+        });
+    if (config.useActivityCosts) {
+        if (!workflow.hitTargets.size) return;
+        bonus.withDefaultCosts().initialize(workflow);
+        if (!DamageBonus.CheckCost(bonus)) return;
+    } else
+        bonus.initialize(workflow);
+    return bonus;
 }
 export const damageBonusToOneRoll = {
     rules: 'all',
@@ -80,6 +90,14 @@ export const damageBonusToOneRoll = {
             type: 'text',
             label: 'CHRISPREMADES.Config.DamageBonus',
             category: 'behavior'
+        },
+        bonusDamageType: {
+            default: [],
+            type: 'select-many',
+            category: 'behavior',
+            label: 'CHRISPREMADES.Macros.Generic.DamageBonusToOneRoll.BonusDamageType',
+            hint: 'CHRISPREMADES.Macros.Generic.DamageBonusToOneRoll.BonusDamageTypeHint',
+            get options() { return constants.damageTypeOptions(); }
         },
         damageType: {
             default: [],
@@ -124,6 +142,14 @@ export const damageBonusToOneRoll = {
             category: 'behavior',
             label:'CHRISPREMADES.Macros.Generic.Common.RollItem'
         },
+        spellLevel: {
+            default: [],
+            type: 'select-many',
+            category: 'behavior',
+            label: 'CHRISPREMADES.Config.SpellLevel',
+            hint: 'CHRISPREMADES.Macros.Generic.Common.SpellLevelHint',
+            get options() { return constants.spellSlotOptions(); }
+        },
         spellSchool: {
             default: [],
             type: 'select-many',
@@ -131,6 +157,19 @@ export const damageBonusToOneRoll = {
             label: 'CHRISPREMADES.Config.SpellSchool',
             hint: 'CHRISPREMADES.Macros.Generic.Common.SpellSchoolHint',
             get options() { return constants.spellSchoolOptions(); }
+        },
+        useActivityCosts: {
+            default: false,
+            type: 'checkbox',
+            category: 'behavior',
+            label:'CHRISPREMADES.Macros.Generic.DamageBonusToOneRoll.Costs'
+        },
+        allowCritical: {
+            default: true,
+            type: 'checkbox',
+            category: 'behavior',
+            label:'CHRISPREMADES.Macros.Generic.DamageBonusToOneRoll.Critical',
+            hint:'CHRISPREMADES.Macros.Generic.DamageBonusToOneRoll.CriticalHint'
         }
     }
 };